Set the default subject in the module.
By setting the default subject with `included` is sets the `subject`
helper on the internal RSpec memoized helper modules. This causes
warnings when another top-level `subject` is defined by the spec author.

This removes using the RSpec memoized helper module and simply sets a
`subject` method delegate for `generator`. The generator is already
memoized by ammeter so we are not missing that feature from RSpec.